perf: serve /en and /es as static (move root layout under [lang])
Move the root layout into app/[lang]/layout.tsx so <html lang> derives from
params.lang instead of headers(), letting both locale routes prerender as static
(SSG) — eliminating per-request rendering, cold starts and the ~1.7s TTFB.
Delete app/layout.tsx and app/page.tsx; the proxy handles the / -> /en redirect.
Theme-flash script, metadata, JSON-LD and per-locale <html lang> all preserved
(verified /en->en, /es->es, / ->307). Resolves PERF-01.
